Add new experimental middle implementations: db
This commits adds a new middle implementations 'db' which is similar to
'pgsql'. All data is stored in the database. A new table structure with
JSONB for tags and relation members is used.

This is EXPERIMENTAL! Not well tested, not complete and performance
might be horrible. It is intended to start a discussion about what we
need and how to best implement it.

To enable these, set the environment variable OSM2PGSQL_MIDDLE to either
'db' or 'new'. If set to 'db' use of the new middle is forced,
independent of the '--slim' setting! Or you can set it to 'new' in which
case the 'db' middle will be used instead of the 'pgsql' middle if and
when that would have been used.

This way of setting the middle is temporary and gives us flexibility
while experimenting, it will not be available in this form in the final
implementation.

The new middle has various options that can be set with the
OSM2PGSQL_MIDDLE_OPTS environment variable. Use a comma separated list
of options, a leading '-' disables the option.

Options:
* drop_tables
* forward_dependencies
* has_bucket_index
* tags
* attributes
* untagged_nodes
* locations
* way_nodes
* relation_members